Fix pasted pictures not shown, when redmine accessed with different domain/host name
This fixes an issue that pasted images are not shown when redmine is accessed with different domain/host name. The issue was that they are inserted into the html with an absolute url (incl. hostname) but they should have been inserted with an relative path.

This PR fixes the issue.